Add dump:upload command with --reset option
This requires us to setup `~/.ssh/config` such that the user account
running this script can push to the `aic-data` repo.

Running this after `dump:export` will push changes to GitHub. If the
`--reset` option is used, the repo will be reset to its initial commit,
and the changes will be force-pushed. This is useful for development,
but we should also consider doing this regularly, maybe every month
or so, to ensure that the git repository doesn't get too bloated.
We'll need to document the `git reset --hard origin/master` workflow.

Also, we should add code to split the CSV files if they get too big.
Set limit to <50 MB or a reasonable amount of lines (~50K).